dead-mans-switch/requirements.txt
python-telegram-bot==22.0
cryptography==44.0.3
flask==3.0.2
flask-cors==4.0.0 